move ApplyToJobButton into Company/ui/
